refactor(payment): avoid ENV access, service pattern, cancel sub
