+++ studio/.env.local.example
@@ -1,5 +1,4 @@
 NEXT_PUBLIC_COSMO_CP_URL=http://localhost:3001
 NEXT_PUBLIC_COSMO_STUDIO_URL=http://localhost:3000
 NEXT_PUBLIC_SLACK_CLIENT_ID=
-
-# NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=
+NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=
